Re-organise types::test_utils dir
