+++ eth2/types/src/test_utils/test_random/public_key.rs
